+++ packages/sdks/src/components/render-content/render-content.lite.tsx
@@ -274,6 +274,7 @@ export default function RenderContent(props: RenderContentProps) {
     setState: state.setContextState,
     context: props.context || {},
     apiKey: props.apiKey,
+    apiVersion: props.apiVersion,
     registeredComponents: state.allRegisteredComponents,
   });
 
@@ -341,6 +342,7 @@ export default function RenderContent(props: RenderContentProps) {
           getContent({
             model: props.model,
             apiKey: props.apiKey,
+            apiVersion: props.apiVersion,
           }).then((content) => {
             if (content) {
               state.mergeNewContent(content);
@@ -354,6 +356,12 @@ export default function RenderContent(props: RenderContentProps) {
       state.emitStateUpdate();
     }
   });
+
+  onUpdate(() => {
+    if (props.content) {
+      state.mergeNewContent(props.content);
+    }
+  }, [props.content]);
 
   onUpdate(() => {
     state.evaluateJsCode();