+++ packages/sdks/typings/augmentations.d.ts
@@ -1,0 +1,12 @@
+import type { JSX as JSXType } from '@builder.io/mitosis/jsx-runtime';
+
+declare module '@builder.io/mitosis/jsx-runtime' {
+  declare namespace JSX {
+    declare interface CustomAttributes extends JSXType.CustomAttributes {
+      /**
+       * Needed to provide data attributes to React Native components
+       */
+      dataSet?: Record<string, string>;
+    }
+  }
+}