+++ packages/sdks/output/react-native/CHANGELOG.md
@@ -1,3 +1,21 @@
+### 0.1.15
+
+- Fix: Removed Columns `width` style causing crash due to `calc()` CSS value.
+
+### 0.1.14
+
+- Fix: Columns styles causing crash due to `number` values being written as `string`s.
+
+### 0.1.13
+
+- Fix: Columns block styling & `stackColumnsAt` prop
+
+### 0.1.12
+
+- Fix: `RenderContent` not re-rendering when its `content` updates. This fixes Symbol rendering in the visual editor.
+- Feature: `RenderContent` and `getContent` now have an `apiVersion` field that can be set to `v2` or `v3` to target specific Builder API versions. Current default is `v2`
+- Fix: Visual Editor drag'n'drop experience
+
 ### 0.1.11
 
 - Fix: Builder blocks no longer add a redundant `Text` wrapper around their `children`.